feat: Main화면 합치기완료. issue 발생 #44, #45
