+++ app/Public Opinion for Stock/Public Opinion for Stock/Assets.xcassets/Color/AccentColor.colorset/Contents.json
@@ -1,15 +1,6 @@
 {
   "colors" : [
     {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0x1E",
-          "green" : "0xB5",
-          "red" : "0x11"
-        }
-      },
       "idiom" : "universal"
     }
   ],
