+++ app/Public Opinion for Stock/Public Opinion for Stock/Models/FavoriteModel.swift
