+++ app/Public Opinion for Stock/Public Opinion for Stock/Presentation/MainView/MainView.swift
@@ -1,10 +1,17 @@
 import SwiftUI
 
 struct MainView: View {
+    @StateObject private var favoriteViewModel = FavoriteViewModel()
+    
     var body: some View {
-        VStack {
-            ListView()
-            ListView()
+        NavigationStack {
+            ScrollView {
+                VStack {
+                    Spacer()
+                    TreeMapView(favoriteViewModel: favoriteViewModel)
+                    ListView()
+                }
+            }
         }
     }
 }
